+++ tools/auth.js
@@ -10,8 +10,6 @@ const {
 	INVALID_TOKEN,
 	MISSING_HEADER,
 	DEACTIVATED_USER,
-	INVALID_CAPTCHA,
-	INVALID_PASSWORD,
 	TOKEN_NOT_FOUND,
 	TOKEN_REVOKED,
 	MULTIPLE_API_KEY,
@@ -22,15 +20,10 @@ const {
 	API_KEY_INVALID,
 	API_KEY_EXPIRED,
 	API_KEY_OUT_OF_SCOPE,
-	API_SIGNATURE_INVALID,
-	CODE_NOT_FOUND,
-	CODE_USED,
-	SAME_PASSWORD
+	API_SIGNATURE_INVALID
 } = require('../messages');
 const { SERVER_PATH } = require('../constants');
 const {
-	NODE_ENV,
-	CAPTCHA_ENDPOINT,
 	BASE_SCOPES,
 	ROLES,
 	ISSUER,
@@ -39,16 +32,11 @@ const {
 	HMAC_TOKEN_EXPIRY,
 	HMAC_TOKEN_KEY
 } = require(`${SERVER_PATH}/constants`);
-const rp = require('request-promise');
 const { getKitSecrets, getFrozenUsers, getNetworkKeySecret } = require('./common');
 const dbQuery = require('./database/query');
-const bcrypt = require('bcryptjs');
 const { getModel } = require('./database/model');
 const { client } = require('./database/redis');
-const uuid = require('uuid/v4');
-const { all, resolve, reject, promisify } = require('bluebird');
-const { sendEmail } = require(`${SERVER_PATH}/mail`);
-const { MAILTYPE } = require(`${SERVER_PATH}/mail/strings`);
+const { reject, promisify } = require('bluebird');
 const { loggerAuth } = require(`${SERVER_PATH}/config/logger`);
 const moment = require('moment');
 const { checkUserOtpActive } = require('./otp');
@@ -394,37 +382,6 @@ const userIsDeactivated = (deactivatedUsers, userId) => {
 	}
 };
 
-const checkCaptcha = (captcha = '', remoteip = '') => {
-	if (!captcha) {
-		if (NODE_ENV === 'development') {
-			return resolve();
-		} else {
-			return reject(new Error(INVALID_CAPTCHA));
-		}
-	} else if (!getKitSecrets().captcha || !getKitSecrets().captcha.secret_key) {
-		return resolve();
-	}
-
-	const options = {
-		method: 'POST',
-		form: {
-			secret: getKitSecrets().captcha.secret_key,
-			response: captcha,
-			remoteip
-		},
-		uri: CAPTCHA_ENDPOINT
-	};
-
-	return rp(options)
-		.then((response) => JSON.parse(response))
-		.then((response) => {
-			if (!response.success) {
-				throw new Error(INVALID_CAPTCHA);
-			}
-			return;
-		});
-};
-
 const checkAdminIp = (whiteListedIps = [], ip = '') => {
 	if (whiteListedIps.length === 0) {
 		return true; // no ip restriction for admin
@@ -482,98 +439,6 @@ const issueToken = (
 		}
 	);
 	return token;
-};
-
-const validatePassword = (userPassword, inputPassword) => {
-	return bcrypt.compare(inputPassword, userPassword);
-};
-
-const getResetPasswordCode = (code) => {
-	return dbQuery.findOne('reset password code', { where: { code } })
-		.then((code) => {
-			if (!code) {
-				const error = new Error(CODE_NOT_FOUND);
-				error.status = 404;
-				throw error;
-			}
-			return code;
-		});
-};
-
-const createResetPasswordCode = (userId) => {
-	return dbQuery.findOne('reset password code', {
-		where: { user_id: userId, used: false },
-		attributes: ['code']
-	})
-		.then((code) => {
-			if (code) {
-				return code;
-			}
-			return getModel('reset password code').create({
-				user_id: userId,
-				code: uuid()
-			});
-		})
-		.then((code) => code.code);
-};
-
-const sendResetPasswordCode = (email, captcha, ip, domain) => {
-	return require('./user').getUserByEmail(email)
-		.then((user) => {
-			return all([ createResetPasswordCode(user.id), user, checkCaptcha(captcha, ip) ]);
-		})
-		.then(([ code, user ]) => {
-			sendEmail(
-				MAILTYPE.RESET_PASSWORD,
-				email,
-				{ code, ip },
-				user.settings,
-				domain
-			);
-			return;
-		});
-};
-
-const isValidPassword = (value) => {
-	return /^(?=.*[a-zA-Z])(?=.*\d).{8,}$/.test(value);
-};
-
-const resetUserPassword = (resetPasswordCode, newPassword) => {
-	if (!isValidPassword(newPassword)) {
-		return reject(new Error(INVALID_PASSWORD));
-	}
-	return getResetPasswordCode(resetPasswordCode)
-		.then((code) => {
-			if (code.used) {
-				throw new Error(CODE_USED);
-			}
-			return code.update({ used: true }, { fields: ['used'] });
-		})
-		.then((code) => require('./user').getUserByKitId(code.user_id, false))
-		.then((user) => user.update({ password: newPassword }, { fields: ['password'] }));
-};
-
-const changeUserPassword = (email, oldPassword, newPassword) => {
-	if (oldPassword === newPassword) {
-		return reject(new Error(SAME_PASSWORD));
-	}
-	if (!isValidPassword(newPassword)) {
-		return reject(new Error(INVALID_PASSWORD));
-	}
-	return require('./user').getUserByEmail(email, false)
-		.then((user) => {
-			return all([ user, validatePassword(user.password, oldPassword) ]);
-		})
-		.then(([ user, passwordIsValid ]) => {
-			if (!passwordIsValid) {
-				throw new Error(INVALID_PASSWORD);
-			} else {
-				return user;
-			}
-		})
-		.then((user) => {
-			return user.update({ password: newPassword });
-		});
 };
 
 const createHmacSignature = (secret, verb, path, expires, data = '') => {
@@ -750,15 +615,9 @@ module.exports = {
 	verifyHmacTokenMiddleware,
 	verifyNetworkHmacToken,
 	userScopeIsValid,
-	isValidPassword,
 	userIsDeactivated,
-	checkCaptcha,
 	findToken,
 	issueToken,
-	validatePassword,
-	sendResetPasswordCode,
-	resetUserPassword,
-	changeUserPassword,
 	getUserKitHmacTokens,
 	createUserKitHmacToken,
 	deleteUserKitHmacToken,
